+++ private/plugins/filemgmt/language/english_utf-8.php
@@ -121,7 +121,29 @@ $LANG_FILEMGMT = array(
     'all_html_allowed' => 'All HTML is allowed',
     'no_new_files'  => 'No new files',
     'no_comments'   => 'No new comments',
-    'more'          => '<em>more ...</em>'
+    'more'          => '<em>more ...</em>',
+    'newly_uploaded' => 'Newly Uploaded',
+    'click_to_view' => 'Click here to view',
+    'no_file_uploaded' => 'No File Uploaded',
+    'description' => 'Description',
+    'category' => 'Category',
+    'err_req_fields' => 'Some required fields were not supplied',
+    'go_back' => 'Go Back',
+    'err_demomode' => 'Uploads are disabled in demo mode',
+);
+
+$LANG_FILEMGMT_ERRORS = array(
+    "1101" => "Upload approval Error: The temporary file was not found. Check error.log",
+    "1102" => "Upload submit Error: The temporary filestore file was not created. Check error.log",
+    "1103" => "The download info you provided is already in the database!",
+    "1104" => "The download info was not complete - Need to enter a title for the new file",
+    "1105" => "The download info was not complete - Need to enter a description for the new file",
+    "1106" => "Upload Add Error: The new file was not created. Check error.log",
+    "1107" => "Upload Add Error: The temporary file was not found. Check error.log",
+    "1108" => "Duplicate file - already existing in filestore",
+    "1109" => "File type not allowed",
+    "1110" => "You must define and select a category for the uploaded file",
+    "9999" => "Unknown Error"
 );
 
 $LANG_FILEMGMT_AUTOTAG = array(
@@ -163,6 +185,8 @@ $LANG_confignames['filemgmt'] = array(
     'enable_rating'         => 'Enable Ratings',
     'displayblocks'         => 'Display glFusion Blocks',
     'silent_edit_default'   => 'Silent Edit Default',
+    'extensions_map'        => 'Extensions used for downloads',
+    'EmailOption'           => 'Email submitter upon approval?',
 );
 $LANG_configsubgroups['filemgmt'] = array(
     'sg_main'               => 'Main Settings'
@@ -205,6 +229,7 @@ define("_MD_SEARCH","Search");
 define("_MD_MAIN","Main");
 define("_MD_SUBMITFILE","Submit File");
 define("_MD_POPULAR","Popular");
+define("_MD_POP", "Pop");   // abbrevision for listing badge
 define("_MD_NEW","New");
 define("_MD_TOPRATED","Top Rated");
 
@@ -281,7 +306,7 @@ define("_MD_INTFILEAT","Interesting Download File at %s"); // %s is your site na
 define("_MD_INTFILEFOUND","Here is an interesting download file I have found at %s"); // %s is your site name
 
 define("_MD_RECEIVED","We received your download information. Thanks!");
-define("_MD_WHENAPPROVED","You'll receive an E-mail when it's approved.");
+define("_MD_WHENAPPROVED","You&apos;ll receive an E-mail when it&apos;s approved.");
 define("_MD_SUBMITONCE","Submit your file/script only once.");
 define("_MD_APPROVED", "Your file has been approved");
 define("_MD_ALLPENDING","All file/script information are posted pending verification.");
@@ -393,6 +418,7 @@ define("_MD_ERRORDESC","ERROR: You need to enter DESCRIPTION!");
 define("_MD_NEWDLADDED","New download added to the database.");
 define("_MD_NEWDLADDED_DUPFILE","Warning: Duplicate File. New download added to the database.");
 define("_MD_NEWDLADDED_DUPSNAP","Warning: Duplicate Snap. New download added to the database.");
+define("_MD_DLUPDATED", "File has been updated.");
 define("_MD_HELLO","Hello %s");
 define("_MD_WEAPPROVED","We approved your download submission to our downloads section. The file name is: ");
 define("_MD_THANKSSUBMIT","Thanks for your submission!");
@@ -438,4 +464,3 @@ define("_MD_COMMENTSWANTED","Comments are appreciated");
 define("_MD_CLICK2SEE","Click to see: ");
 define("_MD_CLICK2DL","Click to download: ");
 define("_MD_ORDERBY","Order By: ");
-?>