+++ private/plugins/filemgmt/language/spanish_colombia_utf-8.php
@@ -121,7 +121,29 @@ $LANG_FILEMGMT = array(
     'all_html_allowed' => 'Todo HTML es permitido',
     'no_new_files'  => 'No hay archivos nuevos',
     'no_comments'   => 'No hay comentarios nuevos',
-    'more'          => '<em>más ...</em>'
+    'more'          => '<em>más ...</em>',
+    'newly_uploaded' => 'Newly Uploaded',
+    'click_to_view' => 'Click here to view',
+    'no_file_uploaded' => 'No File Uploaded',
+    'description' => 'Description',
+    'category' => 'Category',
+    'err_req_fields' => 'Some required fields were not supplied',
+    'go_back' => 'Go Back',
+    'err_demomode' => 'Uploads are disabled in demo mode',
+);
+
+$LANG_FILEMGMT_ERRORS = array(
+    "1101" => "Upload approval Error: The temporary file was not found. Check error.log",
+    "1102" => "Upload submit Error: The temporary filestore file was not created. Check error.log",
+    "1103" => "The download info you provided is already in the database!",
+    "1104" => "The download info was not complete - Need to enter a title for the new file",
+    "1105" => "The download info was not complete - Need to enter a description for the new file",
+    "1106" => "Upload Add Error: The new file was not created. Check error.log",
+    "1107" => "Upload Add Error: The temporary file was not found. Check error.log",
+    "1108" => "Duplicate file - already existing in filestore",
+    "1109" => "File type not allowed",
+    "1110" => "You must define and select a category for the uploaded file",
+    "9999" => "Unknown Error"
 );
 
 $LANG_FILEMGMT_AUTOTAG = array(
@@ -163,6 +185,8 @@ $LANG_confignames['filemgmt'] = array(
     'enable_rating'         => 'Habilitar Calificaciones',
     'displayblocks'         => 'Mostrar los Bloques glFusion',
     'silent_edit_default'   => 'Silent Edit Default',
+    'extensions_map'        => 'Extensiones utilizadas para descargas',
+    'EmailOption'           => '¿Remitente de correo electrónico después de la aprobación?',
 );
 $LANG_configsubgroups['filemgmt'] = array(
     'sg_main'               => 'Main Settings'
@@ -206,6 +230,7 @@ define("_MD_MAIN","Principal");
 define("_MD_SUBMITFILE","Enviar archivo");
 define("_MD_POPULAR","Popular");
 define("_MD_NEW","Nuevo");
+define("_MD_POP", "Pop");   // abbrevision for listing badge
 define("_MD_TOPRATED","Top rankeado");
 
 define("_MD_NEWTHISWEEK","Nuevo esta semana");